requirements.txt
numpy==1.23.5
streamlit==1.28.1
# opencv-python-headless
librosa==0.10.1
tensorflow==2.12.0
matplotlib==3.7.1
Pillow==9.4.0
